Create Arc System Works - Merged Light - Guilty Gear Edition.shader
So Guilty Gear uses a completely different color channels from the ILM and Vertex Colors, as well as different lighting variables so I had to make a separate version of the shader specifically for Guilty Gear. I tested specifically with a Guilty Gear Xrd model, but I assume that both Xrd and SIGN use the same information.