NEW PRESETS FUNCTION
Save and load presets now.

YOU MUST COMPLETELY REINSTALL THE SHADER FOR THIS TO WORK.
Sorry, but due to the required folder structure, I've moved the shader files to its own separate folder in the Assets directory. This means that you will have to completely delete the old "/Assets/Shaders/Aerthas-Unity-Shaders" folder entirely and then reinstall.
None of your materials that exist outside of this folder structure will change.